Lv.2 수정 및 삭제기능 구현

+++ src/main/java/com/example/calender/service/ScheduleServiceImpl.java
@@ -2,9 +2,11 @@ package com.example.calender.service;
 
 import com.example.calender.dto.*;
 import com.example.calender.entity.Schedule;
-import com.example.calender.repository.ScheduleRepository;
+import com.example.calender.entity.User;
+import com.example.calender.repository.*;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.stream.*;
 import java.util.*;
@@ -12,49 +14,90 @@ import java.util.*;
 @Service
 public class ScheduleServiceImpl implements ScheduleService{
     private final ScheduleRepository scheduleRepository;
+    private final UserRepository userRepository;
 
-    public ScheduleServiceImpl(ScheduleRepository scheduleRepository){
+    public ScheduleServiceImpl(ScheduleRepository scheduleRepository, UserRepository userRepository){
         this.scheduleRepository = scheduleRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
     public ScheduleResponse saveSchedule(ScheduleRequest scheduleRequest){
+        User user = userRepository.findByUserName(scheduleRequest.getUserName());
+        if(user == null){
+            System.out.println("신규 등록: " + scheduleRequest.getUserName() + scheduleRequest.getEmail() + scheduleRequest.getPassword());
+            user = new User(scheduleRequest.getUserName(), scheduleRequest.getEmail(), scheduleRequest.getPassword());
+            System.out.println("생성 객체: " + user.getUserName() + user.getEmail() + user.getPassword());
+            userRepository.saveUser(user);
+        }
+        System.out.println("userId : " + user.getUserId());
         LocalDateTime now = LocalDateTime.now();
-        Schedule schedule = new Schedule(scheduleRequest.getContent(), scheduleRequest.getUserId(), now);
+        Schedule schedule = new Schedule(scheduleRequest.getContent(), user.getUserId(), now);
 
-        Schedule savedSchedule = scheduleRepository.saveSchedule(schedule);
+        Schedule savedSchedule = scheduleRepository.saveSchedule(schedule, scheduleRequest.getUserName(), scheduleRequest.getPassword(), scheduleRequest.getEmail());
 
-        return new ScheduleResponse(savedSchedule);
+        String userName = userRepository.findUserNameById(savedSchedule.getUserId());
+        return new ScheduleResponse(savedSchedule, userName);
     }
 
     @Override
     public List<ScheduleResponse> getAllSchedules(){
         List<Schedule> schedules = scheduleRepository.findAll();
-        return schedules.stream().map(ScheduleResponse::new).collect(Collectors.toList());
+        return schedules.stream().map(schedule -> {
+            User user = userRepository.findById(schedule.getUserId());
+            String userName = (user != null) ? user.getUserName() : "Unknown";
+            return new ScheduleResponse(schedule, userName);
+        }).collect(Collectors.toList());
     }
 
     @Override
     public ScheduleResponse getScheduleById(Long id){
         Schedule schedule = scheduleRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("찾을 수 없는 일정입니다."));
-        return new ScheduleResponse(schedule);
+
+        String userName = userRepository.findUserNameById(schedule.getUserId());
+        return new ScheduleResponse(schedule, userName);
     }
 
-    /*@Override
-    public ScheduleResponse updateSchedule(Long id, String task, String userName, String password){
-        Schedule schedule = scheduleRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("수정할 수 없습니다."));
-        schedule.update(task, userName, password);
-        Schedule updatedSchedule = scheduleRepository.saveSchedule(schedule);
-
-        return new ScheduleResponse(updatedSchedule);
-    }*/
-
     @Override
-    public void deleteSchedule(Long id){
+    public void deleteSchedule(Long id, String password){
         Schedule schedule = scheduleRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("제거할 수 없습니다."));
 
+        boolean isCorrect = scheduleRepository.checkPassword(schedule.getUserId(), password);
+        if(!isCorrect){
+            throw new RuntimeException("비밀번호 불일치");
+        }
         scheduleRepository.deleteSchedule(schedule);
+    }
+
+    @Override
+    public List<ScheduleResponse> getFilteredSchedules(Long userId, LocalDate date){
+        List<Schedule> schedules = scheduleRepository.findByFilter(userId, date);
+        return schedules.stream().map(schedule -> {
+            User user = userRepository.findById(schedule.getUserId());
+            String userName = (user != null) ? user.getUserName() : "Unknown";
+            return new ScheduleResponse(schedule, userName);
+        }).collect(Collectors.toList());
+    }
+
+    @Override
+    public ScheduleResponse updateSchedule(Long contentId, String newContent, String newUserName, String password){
+        Schedule schedule = scheduleRepository.findById(contentId)
+                .orElseThrow(() -> new RuntimeException("일정 없음"));
+
+        User existingUser = userRepository.findById(schedule.getUserId());
+        if(existingUser == null || !existingUser.getPassword().equals(password)){
+            throw new RuntimeException("비밀번호 불일치");
+        }
+
+        schedule.setContent(newContent);
+        schedule.setModifiedTime(LocalDateTime.now());
+
+        scheduleRepository.updateSchedule(schedule);
+
+        userRepository.updateUserName(schedule.getUserId(), newUserName);
+
+        return new ScheduleResponse(schedule, newUserName);
     }
 }
